pyproject.toml
[project]
name = "barista-ci-tool"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "en-core-web-sm",
    "fastapi>=0.135.1",
    "fpdf2>=2.8.7",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.16",
    "langchain-groq>=1.1.2",
    "langchain-ollama>=1.0.1",
    "langgraph>=1.0.10",
    "openai>=2.24.0",
    "playwright>=1.58.0",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "scikit-learn>=1.8.0",
    "spacy>=3.8.11",
    "tavily-python>=0.7.22",
    "tldextract>=5.3.1",
    "torch>=2.10.0",
    "transformers>=5.2.0",
    "typing>=3.10.0.0",
    "uvicorn>=0.42.0",
]

[tool.uv.sources]
en-core-web-sm = { url = "https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl" }
